xong hóa đơn

+++ ManagementSystem/ManagementSystem/Product.cs
@@ -23,7 +23,7 @@ namespace ManagementSystem
         public int ID { get; set; }
         public string DisplayName { get; set; }
         public Nullable<int> Counts { get; set; }
-        public Nullable<double> Price { get; set; }
+        public Nullable<long> Price { get; set; }
         public int ID_Unit { get; set; }
         public int ID_Category { get; set; }
